Add more comments and sanity checks to install script

+++ bin/install-dota.sh
@@ -1,20 +1,25 @@
 #!/bin/sh
 
+# Check if we're running as root
 if [ "$(id -u)" != "0" ]; then
     echo "This script must be run as root"
     exit 1
 fi
 
+# Debian
 if [ -f /etc/debian_version ]; then
     if ! grep '7\.' /etc/debian_version >/dev/null; then
         echo "Unsuported Debian version, please use 7.x."
         exit 1
     fi
     echo "Detected Debian system, installing..."
+
+    # Update all packages and install puppet and git
     apt-get update
     apt-get upgrade -y
     apt-get install -y puppet git
 
+# CentOS/Redhat
 elif [ -f /etc/redhat-release ]; then
     if ! grep ' 6\.' /etc/redhat-release >/dev/null; then
         echo "Unsuported CentOS/RedHat version, please use 6.x."
@@ -22,7 +27,7 @@ elif [ -f /etc/redhat-release ]; then
     fi
     echo "Detected CentOS/RedHat system, installing..."
 
-    # Add RPM Forge source
+    # Add RPM Forge source (used for htop and other utilities)
     rpm --import http://apt.sw.be/RPM-GPG-KEY.dag.txt
     rpm -i http://pkgs.repoforge.org/rpmforge-release/rpmforge-release-0.5.3-1.el6.rf.`uname -i`.rpm
 
@@ -38,6 +43,21 @@ else
     exit 1
 fi
 
+# Quick sanity check to make sure everything was installed properly
+if ! command -v git >/dev/null 2>&1; then
+    echo "There was an issue installing git..."
+    exit 1
+fi
+if ! command -v puppet >/dev/null 2>&1; then
+    echo "There was an issue installing puppet..."
+    exit 1
+fi
+
+# Remove any existing files in the puppet directory
 rm -rf /etc/puppet; mkdir /etc/puppet
+
+# Clone the latest copy of the configuration into it
 git clone git://github.com/barroncraft/barroncraft-puppet.git /etc/puppet
+
+# Then run the script to configure the server
 puppet apply /etc/puppet/manifests/dotaSetup.pp
